+++ vscode-dotnet-runtime-library/src/index.ts
@@ -7,6 +7,8 @@ export * from './IExtensionContext';
 export * from './IDotnetAcquireContext';
 export * from './IDotnetListVersionsContext';
 export * from './IDotnetUninstallContext';
+export * from './DotnetVersionSpecRequirement';
+export * from './IDotnetFindPathContext';
 export * from './IDotnetAcquireResult';
 export * from './IDotnetEnsureDependenciesContext';
 export * from './IExtensionContext';
@@ -34,6 +36,8 @@ export * from './Utils/VSCodeEnvironment';
 export * from './Utils/IUtilityContext';
 export * from './Utils/WebRequestWorker';
 export * from './Acquisition/DotnetCoreAcquisitionWorker';
+export * from './Acquisition/DotnetConditionValidator';
+export * from './Acquisition/IDotnetConditionValidator';
 export * from './Acquisition/DotnetInstall';
 export * from './Acquisition/IAcquisitionWorkerContext';
 export * from './Acquisition/DirectoryProviderFactory';
